Moving source files to /lib
