+++ lib/sum.op.js
